Add first prototype for pipelined compilation
This commit adds the first prototype for pipelined compilation. Only the
bridges for Scala 2.11, 2.12 and 2.13 have been modified to support this
feature. 2.10 is unsupported.

To avoid macros used and defined in the same build failing to load, we
return empty URLs from the pickle classpath in 2.11, 2.12 and 2.13. This
limitation has to do with the fact that macros require macro class files
to be classloaded, and therefore cannot be supported in a pipelined
compilation setting. So if we find macros in a build, we disable
pipelining only for that project and warn about it (this warning is only
emitted if fatal warnings is not set).