Add prototype for parallel compilation via outlining
Outlining and parallel compilation is supported for all Scala 2.10,
Scala 2.11, Scala 2.12 and Scala 2.13. Details about how it works are
left for a future blog post or commit. This commit message is still work
in progress and should not be considered stable.

To make sure that outlining does the right thing, we report on some
small statistics from the outline process in the parser.