Introduce Pickles/Tasty compatible API for pipeling
Adds `IRStore` and `IR` to the compiler APIs and replaces the clumsy use
of `List[URI]` in the analysis callback and the external API that was
used to propagate IDs to handles in the core of the compiler bridge.

This change has the benefit of pushing the management of the IRs
lifetime to the implementor of `IRStore` and avoiding altogether the
cleanup of the previously compiler-level map that contained all the IRs
created by the compiler runs.