Don't break binary Zinc APIs and add `resetGlobalIRCaches`
Our previous approach adding the IR store the public interfaces breaks
public APIs (for example, the Dotty bridge wouldn't be able to compile).
To still support those bridges that implement build pipelining, we add
the hooks required to the compiler interface and call them reflectively
from AnalyzingCompiler. If the call fails, we silently swallow it and
log a debug message.